+++ locale/hi/index.js
@@ -7,7 +7,9 @@ module.exports = {
 	short: locale.short,
 	narrow: locale.narrow,
 	// Additional styles.
-	'tiny': require('../../locale-more-styles/hi/tiny.json'),
+	// "tiny" is a legacy name of "mini".
+	'mini-time': require('../../locale-more-styles/hi/mini-time.json'),
+	'tiny': require('../../locale-more-styles/hi/mini-time.json'),
 	// Quantifier.
 	quantify: locale.quantify
 }