+++ source/style/index.js
@@ -1,9 +1,0 @@
-// A preset (style) is an object having shape
-// `{ units, gradation, flavour, custom({ elapsed, time, date, now, locale }) }`.
-//
-// `date` parameter of `custom()` is not guaranteed to be set (can be inferred from `time`).
-//
-export { default as timeStyle } from './time'
-export { default as twitterStyle } from './twitter'
-export { default as approximateStyle } from './approximate'
-export { default as defaultStyle } from './default'